Numerous module upgrades.
Upgraded D3 to 3.5.12 and switched to using npm provided version in node_modules instead of including directly with slycat. Upgraded babel from 6 to 7, along with a bunch of related and depended modules. Upgraded webpack and numerous webpack plugins. Commented out ExportNodeModules plugin for webpack because it crashes with Babel version 7.